Moved XXXLink and XXXIOL vues in ./link subfolder.
