+++ src/components/account/NftsTable.vue
@@ -62,7 +62,7 @@
 <script lang="ts">
 
 import {computed, defineComponent, inject, PropType} from 'vue';
-import TokenLink from "@/components/values/TokenLink.vue";
+import TokenLink from "@/components/values/link/TokenLink.vue";
 import {ORUGA_MOBILE_BREAKPOINT} from '@/App.vue';
 import EmptyTable from "@/components/EmptyTable.vue";
 import {routeManager} from "@/router";
