+++ src/components/token/RoyaltyFeeTable.vue
@@ -68,13 +68,13 @@
 <script lang="ts">
 
 import {defineComponent, PropType} from 'vue';
-import AccountLink from "@/components/values/AccountLink.vue";
+import AccountLink from "@/components/values/link/AccountLink.vue";
 import PlainAmount from "@/components/values/PlainAmount.vue";
 import {ORUGA_MOBILE_BREAKPOINT} from "@/App.vue";
 import {FractionAmount} from "@/schemas/HederaSchemas";
 import StringValue from "@/components/values/StringValue.vue";
 import {TokenInfoAnalyzer} from "@/components/token/TokenInfoAnalyzer";
-import TokenLink from "@/components/values/TokenLink.vue";
+import TokenLink from "@/components/values/link/TokenLink.vue";
 import HbarAmount from "@/components/values/HbarAmount.vue";
 
 export default defineComponent({
