+++ src/components/topic/TopicTable.vue
@@ -77,7 +77,7 @@ import BlobValue from "@/components/values/BlobValue.vue";
 import {ORUGA_MOBILE_BREAKPOINT} from '@/App.vue';
 import EmptyTable from "@/components/EmptyTable.vue";
 import {TransactionTableController} from "@/components/transaction/TransactionTableController";
-import TopicIOL from "@/components/values/TopicIOL.vue";
+import TopicIOL from "@/components/values/link/TopicIOL.vue";
 
 export default defineComponent({
   name: 'TopicTable',
