+++ src/components/transfer_graphs/NftDetailsTransferGraph.vue
@@ -112,8 +112,8 @@
 
 <script lang="ts">
 import {defineComponent, inject, PropType, ref, watch} from "vue";
-import AccountLink from "@/components/values/AccountLink.vue";
-import TokenLink from "@/components/values/TokenLink.vue";
+import AccountLink from "@/components/values/link/AccountLink.vue";
+import TokenLink from "@/components/values/link/TokenLink.vue";
 import ArrowSegment from "@/components/transfer_graphs/ArrowSegment.vue";
 import {NFTTransferLayout} from "@/components/transfer_graphs/layout/NFTTransferLayout";
 import {NftTransactionTransfer} from "@/schemas/HederaSchemas";
