+++ src/components/transfer_graphs/RewardTransferGraph.vue
@@ -101,7 +101,7 @@
 <script lang="ts">
 
 import {computed, defineComponent, inject, PropType} from "vue";
-import AccountLink from "@/components/values/AccountLink.vue";
+import AccountLink from "@/components/values/link/AccountLink.vue";
 import ArrowSegment from "@/components/transfer_graphs/ArrowSegment.vue";
 import HbarAmount from "@/components/values/HbarAmount.vue";
 import HbarExtra from "@/components/values/HbarExtra.vue";
