+++ src/components/transfer_graphs/TokenTransferGraphC.vue
@@ -90,7 +90,7 @@
 <script lang="ts">
 
 import {defineComponent, PropType, ref, watch} from "vue";
-import AccountLink from "@/components/values/AccountLink.vue";
+import AccountLink from "@/components/values/link/AccountLink.vue";
 import TokenAmount from "@/components/values/TokenAmount.vue";
 import ArrowSegment from "@/components/transfer_graphs/ArrowSegment.vue";
 import {TokenTransferLayout} from "@/components/transfer_graphs/layout/TokenTransferLayout";
