+++ src/components/values/link/TopicLink.vue
@@ -24,7 +24,21 @@
 
 <template>
 
-  <EntityIOL :entityId="contractId" :label="label"/>
+  <div class="is-inline-block">
+
+    <template v-if="topicRoute === null">
+      <TopicIOL :topic-id="topicId"/>
+    </template>
+
+    <template v-else>
+      <router-link :to="topicRoute">
+        <span class="h-is-hoverable">
+          <TopicIOL :topic-id="topicId"/>
+        </span>
+      </router-link>
+    </template>
+
+  </div>
 
 </template>
 
@@ -34,32 +48,23 @@
 
 <script lang="ts">
 
-import {computed, defineComponent, onBeforeUnmount, onMounted, PropType} from "vue";
-import {LabelByIdCache} from "@/utils/cache/LabelByIdCache";
-import EntityIOL from "@/components/values/EntityIOL.vue";
+import {computed, defineComponent} from "vue";
+import {routeManager} from "@/router";
+import TopicIOL from "@/components/values/link/TopicIOL.vue";
 
 export default defineComponent({
-  name: "ContractIOL",
-  components: {EntityIOL},
+  name: "TopicLink",
+  components: {TopicIOL},
+
   props: {
-    contractId: {
-      type: String as PropType<string | null>,
-      default: null
-    },
+    topicId: String,
   },
+
   setup(props) {
-    const labelLookup = LabelByIdCache.instance.makeLookup(computed(() => props.contractId))
-    onMounted(
-        () => labelLookup.mount()
-    )
-    onBeforeUnmount(
-        () => labelLookup.unmount()
-    )
-    return {
-      label: labelLookup.entity
-    }
+    const topicRoute = computed(() => props.topicId ? routeManager.makeRouteToTopic(props.topicId) : null)
+    return {topicRoute}
   }
-})
+});
 
 </script>
 
@@ -68,3 +73,4 @@ export default defineComponent({
 <!-- --------------------------------------------------------------------------------------------------------------- -->
 
 <style/>
+
