+++ tests/unit/values/ContractLink.spec.ts
@@ -21,7 +21,7 @@
 import {describe, test, expect} from 'vitest'
 import {mount} from "@vue/test-utils"
 import router from "@/router";
-import ContractLink from "@/components/values/ContractLink.vue";
+import ContractLink from "../../../src/components/values/link/ContractLink.vue";
 
 describe("ContractLink.vue", () => {
