+++ tests/unit/values/TopicLink.spec.ts
@@ -20,7 +20,7 @@
 
 import {describe, it, expect} from 'vitest'
 import {mount} from "@vue/test-utils"
-import TopicLink from "@/components/values/TopicLink.vue"
+import TopicLink from "../../../src/components/values/link/TopicLink.vue"
 import router from "@/router";
 
 describe("TopicLink.vue", () => {
